Rename packages in 'spi' directory
