+++ spi/src/main/java/org/opensearch/jobscheduler/spi/LockModel.java
@@ -24,7 +24,7 @@
  * permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.jobscheduler.spi;
+package org.opensearch.jobscheduler.spi;
 
 import org.opensearch.common.Strings;
 import org.opensearch.common.xcontent.ToXContentObject;
