+++ spi/src/main/java/org/opensearch/jobscheduler/spi/ScheduledJobParameter.java
@@ -24,9 +24,9 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.jobscheduler.spi;
+package org.opensearch.jobscheduler.spi;
 
-import com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule.Schedule;
+import org.opensearch.jobscheduler.spi.schedule.Schedule;
 import org.opensearch.common.xcontent.ToXContentObject;
 
 import java.time.Instant;
