+++ spi/src/main/java/org/opensearch/jobscheduler/spi/schedule/CronSchedule.java
@@ -24,7 +24,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule;
+package org.opensearch.jobscheduler.spi.schedule;
 
 import com.cronutils.model.CronType;
 import com.cronutils.model.definition.CronDefinitionBuilder;
