+++ spi/src/main/java/org/opensearch/jobscheduler/spi/schedule/Schedule.java
@@ -24,7 +24,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule;
+package org.opensearch.jobscheduler.spi.schedule;
 
 import org.opensearch.common.collect.Tuple;
 import org.opensearch.common.io.stream.Writeable;
