+++ spi/src/test/java/org/opensearch/jobscheduler/spi/schedule/CronScheduleTests.java
@@ -24,7 +24,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule;
+package org.opensearch.jobscheduler.spi.schedule;
 
 import com.cronutils.model.time.ExecutionTime;
 import org.opensearch.common.collect.Tuple;
