+++ spi/src/test/java/org/opensearch/jobscheduler/spi/schedule/ScheduleParserTests.java
@@ -24,7 +24,7 @@
  *   permissions and limitations under the License.
  */
 
-package com.amazon.opendistroforelasticsearch.jobscheduler.spi.schedule;
+package org.opensearch.jobscheduler.spi.schedule;
 
 import org.opensearch.common.bytes.BytesArray;
 import org.opensearch.common.xcontent.XContentParser;
