+++ src/main/java/org/opensearch/jobscheduler/ScheduledJobProvider.java
@@ -26,8 +26,8 @@
 
 package org.opensearch.jobscheduler;
 
-import com.amazon.opendistroforelasticsearch.jobscheduler.spi.ScheduledJobParser;
-import com.amazon.opendistroforelasticsearch.jobscheduler.spi.ScheduledJobRunner;
+import org.opensearch.jobscheduler.spi.ScheduledJobParser;
+import org.opensearch.jobscheduler.spi.ScheduledJobRunner;
 
 public class ScheduledJobProvider {
     private String jobType;
